requirements.txt
scikit-optimize
numpy
matplotlib
pyyaml
pandas
tqdm
# Note: meep is installed separately